% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cesium.R
\name{cesium}
\alias{cesium}
\title{Create a Cesium globe widget}
\usage{
cesium(
  data = NULL,
  width = NULL,
  height = NULL,
  padding = 0,
  options = cesium_options(),
  elementId = NULL,
  sizingPolicy = cesium_sizing_policy(padding = padding)
)
}
\arguments{
\item{data}{an sf object}

\item{width}{the width of the globe widget}

\item{height}{the height of the globe widget}

\item{padding}{the padding of the globe widget}

\item{options}{the globe options}

\item{elementId}{Use an explicit element ID for the widget
(rather than an automatically generated one).}

\item{sizingPolicy}{htmlwidgets sizing policy object. Defaults to \code{cesiumSizingPolicy()}}
}
\value{
A HTML widget object
}
\description{
Creates a globe widget using \pkg{htmlwidgets}.
The data argument needs only to be specified if you want
to refer to this object in later add_layers calls. It needs to
be an sf object with either POINT, LINESTRING or POLYGON geometries.
}
