% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czml.R
\name{czml}
\alias{czml}
\alias{czml_color}
\alias{czml_string}
\alias{czml_cartesian3}
\alias{czml_bounding_rectangle}
\alias{czml_cartesian2}
\alias{czml_url}
\alias{czml_double}
\alias{czml_integer}
\alias{czml_near_far_scalar}
\alias{czml_dist_display_cond}
\alias{czml_height_reference}
\alias{czml_horizontal_origin}
\alias{czml_vertical_origin}
\alias{czml_label_style}
\alias{czml_stripe_orientation}
\alias{czml_arc_type}
\alias{czml_shadow_mode}
\alias{czml_classification_type}
\title{CZML values}
\source{
\url{https://github.com/AnalyticalGraphicsInc/czml-writer/wiki/}
}
\usage{
czml_color(color = "white", alpha = NULL, timesteps = NULL, interpolation)

czml_string(string = NULL, timesteps = NULL)

czml_cartesian3(x = NULL, y = NULL, z = NULL, timesteps = NULL, interpolation)

czml_bounding_rectangle(
  x = NULL,
  y = NULL,
  width = NULL,
  height = NULL,
  timesteps = NULL,
  interpolation
)

czml_cartesian2(x = NULL, y = NULL, timesteps = NULL, interpolation)

czml_url(url = NULL, timesteps = NULL)

czml_double(value = NULL, timesteps = NULL, interpolation)

czml_integer(value = NULL, timesteps = NULL, interpolation)

czml_near_far_scalar(
  near_dist = NULL,
  near_value = NULL,
  far_dist = NULL,
  far_value = NULL,
  timesteps = NULL,
  interpolation
)

czml_dist_display_cond(
  near_dist = NULL,
  far_dist = NULL,
  timesteps = NULL,
  interpolation
)

czml_height_reference(
  reference = c("RELATIVE_TO_GROUND", "CLAMP_TO_GROUND", "NONE")
)

czml_horizontal_origin(origin = c("CENTER", "LEFT", "RIGHT"))

czml_vertical_origin(origin = c("CENTER", "TOP", "BOTTOM", "BASELINE"))

czml_label_style(style = c("FILL", "OUTLINE", "FILL_AND_OUTLINE"))

czml_stripe_orientation(orientation = c("HORIZONTAL", "VERTICAL"))

czml_arc_type(arc_type = c("GEODESIC", "RHUMB", "NONE"))

czml_shadow_mode(
  shadow_mode = c("DISABLED", "ENABLED", "CAST_ONLY", "RECEIVE_ONLY")
)

czml_classification_type(
  classification_type = c("BOTH", "TERRAIN", "CESIUM_3D_TILE")
)
}
\arguments{
\item{color}{A character matching an R color code or a hex-color string.}

\item{alpha}{A numeric between 0 and 1 indicating the opacity of the color.}

\item{timesteps}{A vector of timesteps equal to the length of input values
for the property}

\item{interpolation}{If not supplied but timesteps are supplied, the values
will automatically be set to intervals. Use interpolation_options() if
instead you would like to interpolate the property between time steps}

\item{string}{A character string}

\item{x}{Value in x direction}

\item{y}{Value in y direction}

\item{z}{Value in z direction, or height}

\item{width}{Size either in pixels or meters}

\item{height}{Size eiterh in pixels or meters}

\item{url}{A charachter vector pointing to an external URL, a base64-encoded
data URI, or a local file path.}

\item{value}{A single numeric or integer value}

\item{near_dist}{The lower distance in meters}

\item{near_value}{The value of the property at \code{near_dist}}

\item{far_dist}{The upper distance in meters}

\item{far_value}{The value of the property at \code{far_dist}}

\item{reference}{The value a property refers to.}

\item{origin}{The horizontal or vertical origin a property refers to}

\item{style}{The style with wich a label is rendered.}

\item{orientation}{The orientation with which stripes are plotted}

\item{arc_type}{The type that determines how lines are plotted.}

\item{shadow_mode}{The mode with which shadows are drawn to and from an entity.}

\item{classification_type}{A value indicating if a classification has
an effect on terrain and/or 3D tiles.}
}
\value{
A list.
}
\description{
Use these functions set CZML-confirming values for properties
}
