% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_formula}
\alias{eval_formula}
\title{Evaluate list members that are formulae, using the map data as the environment
(if provided, otherwise the formula environment)}
\usage{
eval_formula(list, data)
}
\arguments{
\item{list}{with members as formulae}

\item{data}{map data}
}
\description{
Evaluate list members that are formulae, using the map data as the environment
(if provided, otherwise the formula environment)
}
