% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czml.R, R/icons.R
\name{czml_solid_material}
\alias{czml_solid_material}
\alias{czml_arrow_material}
\alias{czml_grid_material}
\alias{czml_image_material}
\alias{czml_stripe_material}
\alias{czml_checkerboard_material}
\alias{czml_outline_material}
\alias{czml_glow_material}
\alias{czml_dash_material}
\alias{fa_icon}
\title{Materials to be used to draw polygons and lines and other geometries}
\usage{
czml_solid_material(color = czml_color("white"))

czml_arrow_material(color = czml_color("white"))

czml_grid_material(
  color = czml_color("white"),
  alpha = 0.1,
  line_count = czml_cartesian2(8, 8),
  line_thickness = czml_cartesian2(1, 1),
  line_offset = czml_cartesian2(0, 0)
)

czml_image_material(
  image_url = czml_url(NULL),
  image_repeat = czml_cartesian2(1, 1),
  color = czml_color("white"),
  image_transparent = FALSE
)

czml_stripe_material(
  orientation = czml_stripe_orientation(),
  even_color = czml_color("white"),
  odd_color = czml_color("black"),
  offset_stripes = czml_double(0),
  repeat_stripes = czml_double(1)
)

czml_checkerboard_material(
  even_color = czml_color("white"),
  odd_color = czml_color("black"),
  repeat_checker = czml_cartesian2(2, 2)
)

czml_outline_material(
  color = czml_color("white"),
  outline_color = czml_color("black"),
  outline_width = czml_double(1)
)

czml_glow_material(
  color = czml_color("white"),
  glow_power = czml_double(0.25),
  taper_power = czml_double(1)
)

czml_dash_material(
  color = czml_color("white"),
  gap_color = czml_color("black"),
  dash_length = czml_double(16),
  dash_pattern = czml_integer(255L)
)

fa_icon(name, ...)
}
\arguments{
\item{color}{a czml_color used for the main color}

\item{alpha}{value to control the opacity of a color}

\item{line_count}{czml_cartesian2 indicating the number of lines
in x and y direction}

\item{line_thickness}{czml_cartesian2 indicating the thickness
of lines both for lines in x and y direction}

\item{line_offset}{czml_cartesian2 to determine the offset between
lines in x and y direction}

\item{image_url}{czml_url indicating an image used to fill the entity}

\item{image_repeat}{czml_cartesian2 indicating the number of times
an image is repeated in x and y direction}

\item{image_transparent}{logical indicating if the image has a transparent
background or not}

\item{orientation}{czml_stripe_orientation indicating the
orientation used to draw stripes}

\item{even_color}{czml_color for even positions}

\item{odd_color}{czml_color for odd positions}

\item{offset_stripes}{czml_double indicating the offset between stripes}

\item{repeat_stripes}{czml_double indicating the number of times
stripes are repeated}

\item{repeat_checker}{czml_double indicating the number of times
a checkerboard is repeated}

\item{outline_color}{czml_color for the outline}

\item{outline_width}{czml_double for the width of an outline}

\item{glow_power}{czml_double value}

\item{taper_power}{czml_double value}

\item{gap_color}{czml_color to color the gaps between dashes}

\item{dash_length}{czml_double indicating the length of dashes}

\item{dash_pattern}{czml_integer representing a 16-bit bitfield determining
which portions along a single dash_length are the dash (1) and which are
the gap (0)}

\item{name}{character, name of the icon understood by \code{fontawsome::fa()}}

\item{...}{other arguments passed to \code{fontawsome::fa()}.}
}
\description{
Note, that not all materials can be used with every geometry type
}
\section{Functions}{
\itemize{
\item \code{czml_solid_material()}: Solid color material

\item \code{czml_arrow_material()}: Arrow material, lines only

\item \code{czml_grid_material()}: Grid material

\item \code{czml_image_material()}: Image material

\item \code{czml_stripe_material()}: Stripe material

\item \code{czml_checkerboard_material()}: Checkerboard material

\item \code{czml_outline_material()}: Outline material

\item \code{czml_glow_material()}: Glow material, lines only

\item \code{czml_dash_material()}: Dash material, lines only

\item \code{fa_icon()}: Helper function to integrate fontawsome icons

}}
